loan-officer-backend/requirements.txt
django==5.1.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-allauth==0.58.2
dj-rest-auth==5.0.1
django-filter==23.4
Pillow==10.1.0 